Refactor lib and parser files
